APIS-5883 - PRE work to mv DevelopersXXX to DeveloperXXX and Developers2XXX to DevelopersXXX
