+++ app/uk/gov/hmrc/gatekeeper/models/Navigation.scala
@@ -30,7 +30,7 @@ case object StaticNavLinks {
   def apply(): Seq[NavLink] = {
     Seq(
       NavLink("Applications", Some(routes.ApplicationController.applicationsPage(None).url)),
-      NavLink("Developers", Some(routes.Developers2Controller.blankDevelopersPage().url)),
+      NavLink("Developers", Some(routes.DevelopersController.blankDevelopersPage().url)),
       NavLink("Email", Some(routes.EmailsController.landing().url)),
       NavLink("API Approvals", Some(routes.DeploymentApprovalController.pendingPage().url)),
       NavLink("XML", Some(routes.XmlServicesController.organisationsSearchPage().url))
