+++ app/uk/gov/hmrc/gatekeeper/views/NewMainView.scala.html
@@ -64,7 +64,7 @@
 @navigation = @{
   Seq(
     NavigationItem(content = Text("Applications"), href = Some(routes.ApplicationController.applicationsPage(None).url)),
-    NavigationItem(content = Text("Developers"), href = Some(routes.Developers2Controller.blankDevelopersPage().url)),
+    NavigationItem(content = Text("Developers"), href = Some(routes.DevelopersController.blankDevelopersPage().url)),
     NavigationItem(content = Text("Email"), href = Some(routes.EmailsController.landing().url)),
     NavigationItem(content = Text("API Approvals"), href = Some(routes.DeploymentApprovalController.pendingPage().url)),
     NavigationItem(content = Text("XML"), href = Some(routes.XmlServicesController.organisationsSearchPage().url))
