+++ app/uk/gov/hmrc/gatekeeper/views/developers/DeleteDeveloperSuccessView.scala.html
@@ -29,7 +29,7 @@
 <div class="">
     <nav class="breadcrumb-nav">
         <ul>
-            <li class="breadcrumb-nav__item"><a href="@{routes.Developers2Controller.blankDevelopersPage}">Developers</a></li>
+            <li class="breadcrumb-nav__item"><a href="@{routes.DevelopersController.blankDevelopersPage}">Developers</a></li>
             <li class="breadcrumb-nav__item breadcrumb-nav__item--trail" aria-hidden="true"></li>
             <li class="breadcrumb-nav__item">Delete developer</li>
         </ul>
@@ -43,6 +43,6 @@
     </div>
     <p>We've sent an email to the developer and any application administrators to let them know.</p>
 
-    <a href="@{routes.Developers2Controller.blankDevelopersPage}" class="button">Finish</a>
+    <a href="@{routes.DevelopersController.blankDevelopersPage}" class="button">Finish</a>
 </div>
 }
