+++ app/uk/gov/hmrc/gatekeeper/views/developers/DeleteDeveloperView.scala.html
@@ -29,7 +29,7 @@
 <article class="content__body full-width">
     <nav class="breadcrumb-nav">
         <ul>
-            <li class="breadcrumb-nav__item"><a href="@{routes.Developers2Controller.blankDevelopersPage}">Developers</a></li>
+            <li class="breadcrumb-nav__item"><a href="@{routes.DevelopersController.blankDevelopersPage}">Developers</a></li>
             <li class="breadcrumb-nav__item breadcrumb-nav__item--trail" aria-hidden="true"></li>
             <li class="breadcrumb-nav__item">Delete developer</li>
         </ul>
@@ -41,12 +41,12 @@
                 <h1>@developer.user.email</h1>
             </header>
 
-            @helper.form(action = routes.DevelopersController.deleteDeveloperAction(UuidIdentifier(developer.user.userId)), 'class -> "form") {
+            @helper.form(action = routes.DeveloperController.deleteDeveloperAction(UuidIdentifier(developer.user.userId)), 'class -> "form") {
             @helper.CSRF.formField
             <h2>Are you sure you want to delete this developer?</h2>
             <div class="form-field form-field--submit">
                 <button id="submit" class="button--alert" type="submit">Delete developer</button>
-                <a id="cancel" href="@routes.DevelopersController.developerPage(UuidIdentifier(developer.user.userId))" class="button--link">Cancel</a>
+                <a id="cancel" href="@routes.DeveloperController.developerPage(UuidIdentifier(developer.user.userId))" class="button--link">Cancel</a>
             </div>
             }
         } else {
@@ -61,7 +61,7 @@
                     <p><a id="app-@{application.id.value}" href="@{routes.ApplicationController.applicationPage(application.id)}">@application.name</a></p>
                 }
             </div>
-            <a id="finish" href="@routes.DevelopersController.developerPage(UuidIdentifier(developer.user.userId))" class="button">Finish</a>
+            <a id="finish" href="@routes.DeveloperController.developerPage(UuidIdentifier(developer.user.userId))" class="button">Finish</a>
         }
     }
 </article>
