+++ app/uk/gov/hmrc/gatekeeper/views/developers/DeveloperDetailsView.scala.html
@@ -128,12 +128,12 @@
         <br>
         @if(loggedInRequest.role.isSuperUser) {
             <a id="delete-developer" href="@{
-                routes.DevelopersController.deleteDeveloperPage(UuidIdentifier(developer.user.userId))
+                routes.DeveloperController.deleteDeveloperPage(UuidIdentifier(developer.user.userId))
             }" class="button--alert text--center flush--right">Delete developer</a>
         }
         @if(developer.mfaEnabled) {
             <a id="remove-2SV" href="@{
-                routes.DevelopersController.removeMfaPage(UuidIdentifier(developer.user.userId))
+                routes.DeveloperController.removeMfaPage(UuidIdentifier(developer.user.userId))
             }" class="button--alert text--center flush--right">Remove 2SV</a>
         }
     </article>
