+++ app/uk/gov/hmrc/gatekeeper/views/developers/RemoveMfaSuccessView.scala.html
@@ -29,7 +29,7 @@
 <div>
     <nav class="breadcrumb-nav">
         <ul>
-            <li class="breadcrumb-nav__item"><a href="@{routes.Developers2Controller.blankDevelopersPage}">Developers</a></li>
+            <li class="breadcrumb-nav__item"><a href="@{routes.DevelopersController.blankDevelopersPage}">Developers</a></li>
             <li class="breadcrumb-nav__item breadcrumb-nav__item--trail" aria-hidden="true"></li>
             <li class="breadcrumb-nav__item">Remove 2SV</li>
         </ul>
@@ -42,6 +42,6 @@
         <h1 class="transaction-banner--heading">2SV removed</h1>
     </div>
 
-    <a id="finish" href="@{routes.Developers2Controller.blankDevelopersPage}" class="button">Finish</a>
+    <a id="finish" href="@{routes.DevelopersController.blankDevelopersPage}" class="button">Finish</a>
 </div>
 }
