+++ app/uk/gov/hmrc/gatekeeper/views/developers/RemoveMfaView.scala.html
@@ -31,7 +31,7 @@
 <article class="content__body full-width">
     <nav class="breadcrumb-nav">
         <ul>
-            <li class="breadcrumb-nav__item"><a href="@{routes.Developers2Controller.blankDevelopersPage}">Developers</a></li>
+            <li class="breadcrumb-nav__item"><a href="@{routes.DevelopersController.blankDevelopersPage}">Developers</a></li>
             <li class="breadcrumb-nav__item breadcrumb-nav__item--trail" aria-hidden="true"></li>
             <li class="breadcrumb-nav__item">Remove 2SV</li>
         </ul>
@@ -41,12 +41,12 @@
         <h1>@developer.user.email</h1>
     </header>
 
-    @helper.form(action = routes.DevelopersController.removeMfaAction(UuidIdentifier(developer.user.userId)), 'class -> "form") {
+    @helper.form(action = routes.DeveloperController.removeMfaAction(UuidIdentifier(developer.user.userId)), 'class -> "form") {
     @helper.CSRF.formField
     <h2>Are you sure you want to remove 2SV for this developer?</h2>
     <div class="form-field form-field--submit">
         <button id="submit" class="button--alert" type="submit">Remove 2SV</button>
-        <a id="cancel" href="@routes.DevelopersController.developerPage(UuidIdentifier(developer.user.userId))" class="button--link">Cancel</a>
+        <a id="cancel" href="@routes.DeveloperController.developerPage(UuidIdentifier(developer.user.userId))" class="button--link">Cancel</a>
     </div>
     }
 </article>
